syntax formatting in Eva now implements word wrapping and indentation.
